requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0
pytest>=7.4.3
httpx>=0.25.0
pytest-asyncio>=0.21.1
aiofiles>=23.2.1
gitpython>=3.1.40
prisma>=0.11.0
